// src/ui/src/api/ImageboardClient.ts
export class PostDto implements IPostDto {
    id!: number;
    signature?: string | undefined;
    isOp!: boolean;
    text?: string | undefined;
    created!: moment.Moment;
    topicId!: number;
    parentId?: number | undefined;
    children?: PostDto[] | undefined;
    attachments?: AttachmentDto[] | undefined;

    constructor(data?: IPostDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.signature = _data["signature"];
            this.isOp = _data["isOp"];
            this.text = _data["text"];
            this.created = _data["created"] ? moment.parseZone(_data["created"].toString()) : <any>undefined;
            this.topicId = _data["topicId"];
            this.parentId = _data["parentId"];
            if (Array.isArray(_data["children"])) {
                this.children = [] as any;
                for (let item of _data["children"])
                    this.children!.push(PostDto.fromJS(item));
            }
            if (Array.isArray(_data["attachments"])) {
                this.attachments = [] as any;
                for (let item of _data["attachments"])
                    this.attachments!.push(AttachmentDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): PostDto {
        data = typeof data === 'object' ? data : {};
        let result = new PostDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["signature"] = this.signature;
        data["isOp"] = this.isOp;
        data["text"] = this.text;
        data["created"] = this.created ? this.created.toISOString(true) : <any>undefined;
        data["topicId"] = this.topicId;
        data["parentId"] = this.parentId;
        if (Array.isArray(this.children)) {
            data["children"] = [];
            for (let item of this.children)
                data["children"].push(item.toJSON());
        }
        if (Array.isArray(this.attachments)) {
            data["attachments"] = [];
            for (let item of this.attachments)
                data["attachments"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IPostDto {
    id: number;
    signature?: string | undefined;
    isOp: boolean;
    text?: string | undefined;
    created: moment.Moment;
    topicId: number;
    parentId?: number | undefined;
    children?: PostDto[] | undefined;
    attachments?: AttachmentDto[] | undefined;
}

export class AttachmentDto implements IAttachmentDto {
    id!: number;
    originalFilename?: string | undefined;
    filename?: string | undefined;
    contentType?: string | undefined;
    size!: number;
    created!: moment.Moment;
    postId!: number;

    constructor(data?: IAttachmentDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.originalFilename = _data["originalFilename"];
            this.filename = _data["filename"];
            this.contentType = _data["contentType"];
            this.size = _data["size"];
            this.created = _data["created"] ? moment.parseZone(_data["created"].toString()) : <any>undefined;
            this.postId = _data["postId"];
        }
    }

    static fromJS(data: any): AttachmentDto {
        data = typeof data === 'object' ? data : {};
        let result = new AttachmentDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["originalFilename"] = this.originalFilename;
        data["filename"] = this.filename;
        data["contentType"] = this.contentType;
        data["size"] = this.size;
        data["created"] = this.created ? this.created.toISOString(true) : <any>undefined;
        data["postId"] = this.postId;
        return data; 
    }
}

export interface IAttachmentDto {
    id: number;
    originalFilename?: string | undefined;
    filename?: string | undefined;
    contentType?: string | undefined;
    size: number;
    created: moment.Moment;
    postId: number;
}